# lms-backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: MOOC Learning Platform API
  description: >
    Đặc tả API cho nền tảng học tập trực tuyến theo mô hình MOOC.
    Bao gồm 3 nhóm tác nhân: Student (Học viên), Instructor (Giảng viên), Admin (Quản trị viên).
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local dev server
...
security:
  - bearerAuth: []
paths:
...
  /sections/{id}:
    put:
      tags: [Course - Instructor]
      summary: Cập nhật chương học
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                orderIndex: { type: integer }
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SectionResponse' }
    delete:
      tags: [Course - Instructor]
      summary: Xóa chương học
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '204':
          description: Xóa thành công
...
  /quizzes/{id}:
    put:
      tags: [Quiz - Instructor]
      summary: Cập nhật thông tin bài kiểm tra
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                passingScore: { type: integer }
                durationMinutes: { type: integer }
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QuizResponse' }
    delete:
      tags: [Quiz - Instructor]
      summary: Xóa bài kiểm tra
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '204':
          description: Xóa thành công
...
  /quizzes/{id}/submit:
    post:
      tags: [Quiz - Student]
      summary: Nộp bài kiểm tra, hệ thống tự động chấm điểm
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [answers]
              properties:
                answers:
                  type: array
                  items:
                    type: object
                    properties:
                      questionId: { type: integer }
                      selectedAnswerId: { type: integer }
      responses:
        '200':
          description: Kết quả chấm điểm tức thì
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QuizResultResponse' }

  # ==========================================================
  # CERTIFICATE
  # ==========================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPathParam:
      name: id
      in: path
      required: true
      schema: { type: integer }
  schemas:
...
    SectionResponse:
      type: object
      properties:
        id: { type: integer }
        courseId: { type: integer }
        title: { type: string }
        orderIndex: { type: integer }
        lessons:
          type: array
          items: { $ref: '#/components/schemas/LessonResponse' }
...
    LessonResponse:
      type: object
      properties:
        id: { type: integer }
        sectionId: { type: integer }
        title: { type: string }
        contentType: { type: string, enum: [Video, Document, Text] }
        contentUrl: { type: string }
        durationMinutes: { type: integer }
        orderIndex: { type: integer }
...
    QuizResponse:
      type: object
      properties:
        id: { type: integer }
        courseId: { type: integer }
        title: { type: string }
        passingScore: { type: integer }
        durationMinutes: { type: integer }
        questions:
          type: array
          items: { $ref: '#/components/schemas/QuestionResponse' }
...
    QuestionResponse:
      type: object
      properties:
        id: { type: integer }
        quizId: { type: integer }
        questionText: { type: string }
        point: { type: integer }
        answers:
          type: array
          items: { $ref: '#/components/schemas/AnswerResponse' }
    AnswerResponse:
      type: object
      properties:
        id: { type: integer }
        questionId: { type: integer }
        answerText: { type: string }
        isCorrect:
          type: boolean
          description: Chỉ trả về cho Instructor/Admin, ẩn với Student khi làm bài
...
    QuizResultResponse:
      type: object
      properties:
        quizId: { type: integer }
        score: { type: number, format: float }
        passingScore: { type: integer }
        isPassed: { type: boolean }
        correctCount: { type: integer }
        totalQuestions: { type: integer }
...